+++ CHANGELOG.md
@@ -17,6 +17,8 @@ resources.
 - Add support for the `--format` flag in the `sensuctl command list` subcommand.
 - Namespace can be ommited from event when performing an HTTP POST request to
 the `/events` endpoint.
+- Fixed a bug where failing check TTL events could occur event if keepalive
+failures had already occurred.
 
 ## [5.16.1] - 2019-12-18